A bit of memory management
DiskShaderCacheUtil went non util as at Runtime.addShutDownHook() exec the gl context is gone, and its mesh and utils can't be properly disposed.
IShaderPipeline went disposable.
IWrappedShader went the React.useEffect route for now to force users to declare disposable, bound resources. The Gdx ShaderProgram api doesn't make it easy to infor this, but there might be a very functional, abstract declaration possible for bindResource that take a method reference of the shader program, and a bindable resource. But that's not going to be easy.