Fully dynamic variable shader
Going pure lwjgl since the LibGDX shader program is too limited to facilitate the shader graph. With this new addition, each of fragment, vertex, and geometry (new), can be compiled, linked and rendered individually or combined. Also allows for runtime updates of each of the 3.
Suitable solution for binding shader specific resources has to be found (again), but itll probably be something akin to inversion of control. Handing the reigns to the individual shader component to do its thing.