melange-shading extracted
With the content of melange-shading exceeding a managable size and compelexity without further visibility restriction or dependency managament (think of JPMS what you will), there's but one solution: Extract it as a separate jar. This was always the backup plan, and remains so for what will be melange-mesh, but mildly unfortunate though managable thanks to the brilliant minds of jitpack.